requirements_infer.txt
opencv-python==4.4.0.46
matplotlib=3.3.3
pandas=1.1.3
pillow==8.4.0
pytorch-lightning==1.4.2
rasterio==1.2.10
scipy=1.5.2
scikit-image=0.17.2
scikit-learn=0.23.2
pytorch=1.9.0
torchvision==0.10.0
xgboost==1.4.2